shared/t4_lambda_shared/utils.py
"""
Helper functions.
"""
import json
import os


def get_default_origins():
    """
    Returns a list of origins that should normally be passed into the @api decorator.
    """
    return [
        'http://localhost:3000',
        os.environ.get('WEB_ORIGIN')
    ]


def make_json_response(status_code, json_object, extra_headers=None):
    """
    Helper function to serialize a JSON object and add the JSON content type header.
    """
    headers = {
        "Content-Type": 'application/json'
    }
    if extra_headers is not None:
        headers.update(extra_headers)

    return status_code, json.dumps(json_object), headers
